api-gateway/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.8.0
python-jose[cryptography]==3.3.0  # Para JWT en auth-service
passlib[bcrypt]==1.7.4  # Para hashear contraseñas
python-dotenv==1.0.1  # Para manejar variables de entorno
httpx==0.27.2  # Para comunicación entre microservicios (en api-gateway)
pydantic==2.1.1  # Para validación de datos
pyjwt==2.6.0  
python-multipart